feat: Use Drawer instead of Dialog for merging cubes
